add snapshot tests and refactor Calculator component
