+++ src/components/home/home.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import Home from './Home';
+
+it('renders correctly', () => {
+  const tree = renderer
+    .create(<Home />)
+    .toJSON();
+  expect(tree).toMatchSnapshot();
+});
